Migrate daemon.rs TmpRepo test to integration test
Move recent_working_log_snapshot_preserves_humans_on_restore to
tests/integration/daemon_unit.rs using TestRepo. Make
RecentWorkingLogSnapshot and restore_recent_working_log_snapshot pub.
