+++ src/daemon.rs
@@ -1884,7 +1884,7 @@ fn capture_recent_working_log_snapshot(
     })))
 }
 
-fn restore_recent_working_log_snapshot(
+pub fn restore_recent_working_log_snapshot(
     repo: &Repository,
     base_commit: &str,
     snapshot: &RecentWorkingLogSnapshot,
@@ -3681,12 +3681,13 @@ struct PendingRootSlot {
 }
 
 #[derive(Debug, Clone, Default)]
-struct RecentWorkingLogSnapshot {
-    files: HashMap<String, Vec<crate::authorship::attribution_tracker::LineAttribution>>,
-    prompts: HashMap<String, crate::authorship::authorship_log::PromptRecord>,
-    file_contents: HashMap<String, String>,
-    humans: std::collections::BTreeMap<String, crate::authorship::authorship_log::HumanRecord>,
-    sessions: std::collections::BTreeMap<String, crate::authorship::authorship_log::SessionRecord>,
+#[doc(hidden)]
+pub struct RecentWorkingLogSnapshot {
+    pub files: HashMap<String, Vec<crate::authorship::attribution_tracker::LineAttribution>>,
+    pub prompts: HashMap<String, crate::authorship::authorship_log::PromptRecord>,
+    pub file_contents: HashMap<String, String>,
+    pub humans: std::collections::BTreeMap<String, crate::authorship::authorship_log::HumanRecord>,
+    pub sessions: std::collections::BTreeMap<String, crate::authorship::authorship_log::SessionRecord>,
 }
 
 impl RecentWorkingLogSnapshot {
@@ -8739,65 +8740,6 @@ mod tests {
         assert_eq!(normalized.get("example.txt"), carryover.get("example.txt"));
     }
 
-    #[test]
-    fn recent_working_log_snapshot_preserves_humans_on_restore() {
-        use crate::authorship::attribution_tracker::LineAttribution;
-        use crate::authorship::authorship_log::HumanRecord;
-        use crate::git::test_utils::TmpRepo;
-        use std::collections::BTreeMap;
-
-        let test_repo = TmpRepo::new().expect("Failed to create test repo");
-        let repo = test_repo.gitai_repo();
-
-        // Create a snapshot with KnownHuman attributions
-        let h_hash = "h_abc123";
-        let human_record = HumanRecord {
-            author: "Test User <test@example.com>".to_string(),
-        };
-
-        let file_path = "test.txt";
-        let line_attributions = vec![LineAttribution {
-            start_line: 1,
-            end_line: 1,
-            author_id: h_hash.to_string(),
-            overrode: None,
-        }];
-
-        let mut humans = BTreeMap::new();
-        humans.insert(h_hash.to_string(), human_record.clone());
-
-        let snapshot = RecentWorkingLogSnapshot {
-            files: HashMap::from([(file_path.to_string(), line_attributions.clone())]),
-            prompts: HashMap::new(),
-            file_contents: HashMap::from([(file_path.to_string(), "test line\n".to_string())]),
-            humans: humans.clone(),
-            sessions: BTreeMap::new(),
-        };
-
-        // Restore the snapshot
-        let base_commit = "HEAD";
-        let restored = restore_recent_working_log_snapshot(repo, base_commit, &snapshot).unwrap();
-        assert!(restored, "Snapshot should be restored");
-
-        // Read back the INITIAL file and verify humans are present
-        let working_log = repo
-            .storage
-            .working_log_for_base_commit(base_commit)
-            .unwrap();
-        let initial = working_log.read_initial_attributions();
-
-        // Verify humans were restored
-        assert_eq!(
-            initial.humans.len(),
-            1,
-            "Should have one human record after restore"
-        );
-        assert_eq!(
-            initial.humans.get(h_hash),
-            Some(&human_record),
-            "Human record should match"
-        );
-    }
 
     // -----------------------------------------------------------------------
     // Readonly command ingress fast-path tests
